trails_service/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-request==1.7.0
pillow==11.1.0
requests==2.32.3